Draw Top Square
