+++ LambdaLogo/LambdaLogo.xcodeproj/project.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
