+++ LambdaLogo/LambdaLogo.xcodeproj/xcuserdata/tobikuyo.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>XOXO Logo.xcscheme_^#shared#^_</key>
+		<key>LambdaLogo.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
