+++ LambdaLogo/LambdaLogo/Assets.xcassets/AppIcon.appiconset/Contents.json
