+++ LambdaLogo/LambdaLogo/Assets.xcassets/Contents.json
