+++ LambdaLogo/LambdaLogo/Info.plist
