+++ LambdaLogo/LambdaLogo/ViewController.swift
@@ -1,0 +1,22 @@
+//
+//  ViewController.swift
+//  LambdaLogo
+//
+//  Created by Tobi Kuyoro on 19/02/2020.
+//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+    
+    @IBOutlet weak var logoView: UIView!
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
