--- XOXO Logo/XOXO Logo/Base.lproj/Main.storyboard
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="15702" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="pcJ-iN-GaW">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="15704"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Logo View Controller-->
-        <scene sceneID="g02-Aj-ifV">
-            <objects>
-                <viewController id="pcJ-iN-GaW" customClass="LogoViewController" customModule="XOXO_Logo" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="VHL-kY-kLc">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="mJ5-oy-Zq2" customClass="LogoView" customModule="XOXO_Logo" customModuleProvider="target">
-                                <rect key="frame" x="7" y="398" width="400" height="100"/>
-                                <subviews>
-                                    <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="TM" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="6gN-0x-YPp">
-                                        <rect key="frame" x="374" y="0.0" width="26" height="21"/>
-                                        <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-                                        <fontDescription key="fontDescription" type="boldSystem" pointSize="15"/>
-                                        <nil key="textColor"/>
-                                        <nil key="highlightedColor"/>
-                                    </label>
-                                </subviews>
-                                <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                                <constraints>
-                                    <constraint firstAttribute="width" constant="400" id="goM-CF-207"/>
-                                    <constraint firstAttribute="height" constant="100" id="mkI-cU-BPf"/>
-                                </constraints>
-                            </view>
-                        </subviews>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                        <constraints>
-                            <constraint firstItem="mJ5-oy-Zq2" firstAttribute="centerY" secondItem="VHL-kY-kLc" secondAttribute="centerY" id="FZP-FU-KoM"/>
-                            <constraint firstItem="mJ5-oy-Zq2" firstAttribute="centerX" secondItem="VHL-kY-kLc" secondAttribute="centerX" id="N9T-to-kHp"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="NfG-wP-Q1c"/>
-                    </view>
-                    <connections>
-                        <outlet property="logoView" destination="mJ5-oy-Zq2" id="lIQ-V6-3WX"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="nhY-O9-KVz" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="113" y="151"/>
-        </scene>
-    </scenes>
-</document>
