--- XOXO Logo/XOXO Logo/LogoViewController.swift
@@ -1,18 +1,0 @@
-//
-//  LogoViewController.swift
-//  XOXO Logo
-//
-//  Created by Tobi Kuyoro on 22/01/2020.
-//  Copyright © 2020 Tobi Kuyoro. All rights reserved.
-//
-
-import UIKit
-
-class LogoViewController: UIViewController {
-    
-    @IBOutlet weak var logoView: LogoView!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-    }
-}
